refactor: split functions into individual files
