+++ src/deep-clone/deep-clone.ts
@@ -1,0 +1,6 @@
+import { deepCloneSync } from './deep-clone-sync';
+import type { DeepCloneOptions } from './typings';
+
+export async function deepClone<T>(target: T, options?: DeepCloneOptions): Promise<T> {
+  return deepCloneSync<T>(target, options);
+}
