+++ src/deep-clone/index.ts
@@ -1,23 +1,2 @@
-import cloneDeep from 'lodash-es/cloneDeep';
-
-export interface DeepCloneOptions {
-  absolute?: boolean;
-}
-
-export function deepCloneSync<T>(
-  target: T,
-  options?: DeepCloneOptions
-): T {
-  if (target == null) throw new TypeError(`'target' is not defined`);
-
-  return options && options.absolute
-    ? cloneDeep<T>(target)
-    : JSON.parse(JSON.stringify(target)) as T;
-}
-
-export async function deepClone<T>(
-  target: T,
-  options?: DeepCloneOptions
-): Promise<T> {
-  return deepCloneSync<T>(target, options);
-}
+export * from './deep-clone';
+export * from './deep-clone-sync';
