+++ src/deep-clone/typings.ts
@@ -1,0 +1,3 @@
+export interface DeepCloneOptions {
+  absolute?: boolean;
+}
